+++ src/handlers/test-handlers-enhanced/index.ts
@@ -4,6 +4,7 @@ import { testTaskOperations } from "./task-tests.js";
 import { testSubtaskOperations } from "./subtask-tests.js";
 import { testLabelOperations } from "./label-tests.js";
 import { testBulkOperations } from "./bulk-tests.js";
+import { testCommentOperations } from "./comment-tests.js";
 import { TestSuite, ComprehensiveTestReport } from "./types.js";
 
 export async function handleTestAllFeaturesEnhanced(
@@ -17,6 +18,7 @@ export async function handleTestAllFeaturesEnhanced(
   suites.push(await testSubtaskOperations(todoistClient));
   suites.push(await testLabelOperations(todoistClient));
   suites.push(await testBulkOperations(todoistClient));
+  suites.push(await testCommentOperations(todoistClient));
 
   // Calculate totals
   const totalTests = suites.reduce((sum, suite) => sum + suite.tests.length, 0);
@@ -53,3 +55,4 @@ export { testTaskOperations } from "./task-tests.js";
 export { testSubtaskOperations } from "./subtask-tests.js";
 export { testLabelOperations } from "./label-tests.js";
 export { testBulkOperations } from "./bulk-tests.js";
+export { testCommentOperations } from "./comment-tests.js";